Add code for simulation
the simulation will run pathFind and draw the maze and path found
it is possible for no path to be found and in that case only the maze is drawn
in order to enable the gui SFML must be setup properly and line 4 of Window.h must
be uncommented

Since there was no base to build the simulation on, both pathfinding and simulation
were developed together. This made for some weird merging merging that would not
follow our workflow. So instead I combined all changes to simulation into this
single commit to maintain the proper workflow. In future development this should
not be necessary.