Created CompactMaze Class
The compact maze class uses two flag matrices to keep track of which
paths are open, closed, and explored. It will have a function to
randomly generate a maze, and will be used by MouseBot as a virtual maze
for testing.

I also changed the Path's nextStep() to popStep() so It will make sense
with the maze generation algorithm, and added a * overload to Pos so it
can be multiplied by a scalar.